+++ Source/Objects/Data Processing/Histogram Objects/Waveform/ORWaveformController.m
@@ -141,10 +141,10 @@
 - (id)tableView:(NSTableView *)tableView objectValueForTableColumn:(NSTableColumn *)tableColumn row:(NSInteger)row
 {
 	if([[tableColumn identifier] isEqualToString:@"Value"]){
-        return [NSNumber numberWithInteger:[model value:row]];
+        return [NSNumber numberWithInteger:[model value:(int)row]];
 	}
     else if([[tableColumn identifier] isEqualToString:@"Hex"]){
-		return [NSString stringWithFormat:@"0x%x",[model value:row]];
+		return [NSString stringWithFormat:@"0x%x",[model value:(int)row]];
 	}
 
 	else return [NSNumber numberWithInteger:row];
